fix bugs

+++ script.js
@@ -1,43 +1,70 @@
+const ratio = .1
+const options = {
+  root: null,
+  rootMargin: '0px',
+  threshold: ratio
+}
+
+const handleIntersect = function (entries, observer) {
+  entries.forEach(function (entry) {
+    if (entry.intersectionRatio > ratio) {
+      entry.target.classList.add('reveal-visible')
+      observer.unobserve(entry.target)
+    }
+  })
+}
+
+const observer = new IntersectionObserver(handleIntersect, options)
+document.querySelectorAll('.reveal').forEach(function(r){
+    observer.observe(r)
+})
+
+
 const resultsTable = document.querySelector("#trainingResults");
 const add = document.querySelector("#createTraining");
 const supp = document.querySelector("#delete");
 
-add.onmouseup = function addTraining(){
-    if(document.getElementById("age").value == "" || document.getElementById("size").value == "" || document.getElementById("weight").value == "" || document.getElementById("userGoal").value == ""){
-        alert("Veuillez entrer toutes vos informations")
-    }else{
-        //First we calculate the imc of the user
-        const imc = IMCcalculator(document.getElementById("size").value, document.getElementById("weight").value);
-        //Then we get the gender the age and the target of the user in order to calculate the better training program
-        const userGender = getGender();
-        const userAge = document.getElementById("age").value;
-        const userGoal = document.getElementById("userGoal").value;
-        //Now we get the selected exercices
-        const currentExercice1 = document.getElementById("exercice1").value;
-        const currentExercice2 = document.getElementById("exercice2").value;
-        const currentExercice3 = document.getElementById("exercice3").value;
-        const currentExercice4 = document.getElementById("exercice4").value;
-        const currentExercice5 = document.getElementById("exercice5").value;
-        const exercicesArray = [currentExercice1, currentExercice2, currentExercice3, currentExercice4, currentExercice5];
-        //Now we create the table new row
-        if(currentExercice1 == "" && currentExercice2 == "" && currentExercice3 == "" && currentExercice4 == "" && currentExercice5 == ""){
-            alert("Veuillez entrer au moins un exercice")
+if(add!= null){
+    add.onmouseup = function addTraining(){
+        if(document.getElementById("age").value == "" || document.getElementById("size").value == "" || document.getElementById("weight").value == "" || document.getElementById("userGoal").value == ""){
+            alert("Veuillez entrer toutes vos informations")
         }else{
-            exercicesArray.forEach(element => {
-                if(element != ""){
-                    addNewRow(element, getRepetition(element, imc, userGender, userAge, userGoal));
-                }
-            });
+            //First we calculate the imc of the user
+            const imc = IMCcalculator(document.getElementById("size").value, document.getElementById("weight").value);
+            //Then we get the gender the age and the target of the user in order to calculate the better training program
+            const userGender = getGender();
+            const userAge = document.getElementById("age").value;
+            const userGoal = document.getElementById("userGoal").value;
+            //Now we get the selected exercices
+            const currentExercice1 = document.getElementById("exercice1").value;
+            const currentExercice2 = document.getElementById("exercice2").value;
+            const currentExercice3 = document.getElementById("exercice3").value;
+            const currentExercice4 = document.getElementById("exercice4").value;
+            const currentExercice5 = document.getElementById("exercice5").value;
+            const exercicesArray = [currentExercice1, currentExercice2, currentExercice3, currentExercice4, currentExercice5];
+            //Now we create the table new row
+            if(currentExercice1 == "" && currentExercice2 == "" && currentExercice3 == "" && currentExercice4 == "" && currentExercice5 == ""){
+                alert("Veuillez entrer au moins un exercice")
+            }else{
+                exercicesArray.forEach(element => {
+                    if(element != ""){
+                        addNewRow(element, getRepetition(element, imc, userGender, userAge, userGoal));
+                    }
+                });
+            }
+            new_lines = true;
         }
-        new_lines = true;
     }
 }
 
-supp.onclick = function deleteAll(){
-    while (resultsTable.rows.length > 2){
-        resultsTable.deleteRow(resultsTable.rows.length-1);
+if(supp != null){
+    supp.onclick = function deleteAll(){
+        while (resultsTable.rows.length > 2){
+            resultsTable.deleteRow(resultsTable.rows.length-1);
+        }
     }
 }
+
 
 function addNewRow(exercice, repetition){
     const newRow = document.createElement("tr");
@@ -192,23 +219,3 @@ function getRepetition(exercice, imc, gender, age, objectif){
 }
 
 
-const ratio = .1
-const options = {
-  root: null,
-  rootMargin: '0px',
-  threshold: ratio
-}
-
-const handleIntersect = function (entries, observer) {
-  entries.forEach(function (entry) {
-    if (entry.intersectionRatio > ratio) {
-      entry.target.classList.add('reveal-visible')
-      observer.unobserve(entry.target)
-    }
-  })
-}
-
-const observer = new IntersectionObserver(handleIntersect, options)
-document.querySelectorAll('.reveal').forEach(function(r){
-    observer.observe(r)
-})
